Fixed: JobConfiguration component opens up below the jobs instead of right side(#2vjxdha)
Currently in process of fixing this bug, jobConfiguration component opens up below instead of right side. Using emitter to emit the job from MoreJobs component to the component it is placed in & when we click on the job in MoreJobs component, it send the job details through emitter so respective component recieves the job details and that component jobConfiguration will open up the configuration in right side. However it doesn't seems to work.